[DEV] client start

+++ NetworkLibrary/NetworkLibrary/Socket.cpp
@@ -93,7 +93,7 @@ void Socket::Send(Message& msg)
 void Socket::Recv()
 {
 	int len = 0;
-	int retval = recvn(m_socket, (char*)&len, sizeof(int), 0);
+	int retval = recvn(m_socket, (char*)&len, sizeof(short), 0);
 	if (retval == SOCKET_ERROR)
 		throw Exception("recv failed");
